Fix canister test

+++ src/internet_identity_interface/src/internet_identity/types/api_v2.rs
@@ -166,7 +166,7 @@ pub enum IdentityPropertiesReplaceError {
         space_required: u64,
     },
     NameTooLong {
-        limit: usize,
+        limit: u64,
     },
     InternalCanisterError(String),
 }
